Adds example files, updates form history etc. DIG-1004
