+++ config/default/node_elections.settings.yml
@@ -1,1 +1,29 @@
-upload_directory: sites/default/files/election_results
+upload_directory: 'public://election_results'
+history:
+  -
+    generate_date: 1636157950
+    upload_date: 1664580711
+    file: 361483
+    result: Failed
+  -
+    generate_date: 1636157950
+    upload_date: 1664810953
+    file: 361485
+    result: Success
+  -
+    generate_date: 1636157950
+    upload_date: 1664821472
+    file: 361486
+    result: Success
+  -
+    generate_date: 1636157950
+    upload_date: 1664911992
+    file: 361489
+    result: Success
+  -
+    generate_date: 1636157950
+    upload_date: 1664922230
+    file: 361493
+    result: Success
+    election: '34858'
+last-run: null
